requirements.txt
flask==3.1.0
pymongo==4.12.0
python-dotenv==1.1.0
flask-cors==5.0.1
gunicorn==22.0.0
dnspython==2.7.0
